cli/pyproject.toml
[project]
name = "fp-cli"
version = "0.1.0"
description = "CLI for the FastAPI Platform — deploy Python APIs from your terminal"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "watchfiles>=0.21.0",
]

[project.scripts]
fp = "fp_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fp_cli"]
